+++ resources/views/hmif/dashboard.blade.php
@@ -1,9 +1,16 @@
 @extends('layouts.starter')
 
 @section('header-content')
-    <h3>HMIF</h3>
+    <h3>Halo Anggota HMIF</h3>
 @endsection
 
 @section('main-content')
-    <h1>Hello World</h1>
+    <div class="container mt-4">
+        <h1>Halo Anggota HMIF, {{ Auth::user()->name }}</h1>
+
+        <!-- Buttons -->
+        <a href="{{ url('hmif/ketersediaanRuangan') }}" class="btn btn-primary">Cek Ketersediaan Ruangan</a>
+        <a href="{{ url('hmif/pengajuanRuangan') }}" class="btn btn-primary mr-2">Ajukan Peminjaman Ruangan</a>
+        <a href="{{ url('hmif/statusPemRuangan') }}" class="btn btn-primary mr-2">Cek Status Peminjaman Ruangan</a>
+    </div>
 @endsection
